Switching to an event model helps decouple the model and the view further. Rather than reaching in to the model whenever the timer fires (and after calling PomodoroModel#tick), the view code in PomodoroTimer registers to be notified when there is an interesting state change. This paves the way for extracting the timer from the GUI.
The tests are all broken, but fixing them now is low-value busywork: I'm not done with the changes to the design that I want to make, so continuously adapting them through each intermediate phase is painful. If I was doing this on production software and would benefit from a safety net, I'd probably have started with GUI tests. All of that has been stable so far, and so all those tests would still be valid at this point. Once the work under the hood is done, then the slow GUI tests could be tossed (or stripped down) in favour of faster unit tests.