Several changes here. First, I group data that travels together, namely the session durations and the current state. This cleans things up a little and adds some semantic meaning. I also introduce more descriptive domain events for the GUI to handle. This simplifies telling the difference between, for example, a session ending or a countdown tick. The GUI is pretty close to being just bindings on the data model: it pulls data from the events and sets them in the appropriate widgets.
I opted to leave the timer alone, despite hinting that pulling it out made sense. The timer itself is arguably not part of the pomodoro "algorithm"; it drives the ticks, but isn't really part of the state transitions themselves. I thought it not worth the added complexity to pull that in to the model, and rather let the UI be responsible for deciding how to drive the model.

I also allow the current state to be passed in via the constructor. This allows the timer to be initialized in whatever state we want, which will greatly aid testing later. It's also a valuable change should we want to persist the current state and resume, say between runs of the application.