Alternative testable implementation of the Pomodoro timer. In this version, the model behaves as a pure function, determining the next state from a given state. Tracking the current state of the application is pulled out to the GUI. Essentially, rather than pushing events out in response to state updates, the model represents each transition (including the end of each session) and relies on the caller to pull the next state when it's ready to do so. The GUI doesn't particularly care about the end of a session, so when it reaches such a state, it simply pulls the next one.
Tests are of similar clarity to the previous version in that they showcase all the state transitions and transparently carry the number of completed pomodoros and remaining duration. Since this is all encoded in the states themselves, special test doubles such as the capturing event listener are no longer required.