The main theme behind the changes was to consider that there are effectively three different types of object in the object graph of this application: 1. Configuration objects, such as the maximum withdrawal 2. Leaf nodes of the object graph, such as the database 3. Dynamic business rules objects, such as the logged-in user commands
I find that how the pieces fit together becomes less of a treasure hunt when objects that fit in categories 1 and 2 are lifted and explicitly wired at the top level. This gives a sort of table of contents of the application, at the point of instantiation; I don't have to go digging through multiple levels of indirection to understand configuration parameters.

Leaf nodes are a little more interesting: the fact that they are at the boundaries means that they should not be relying on any of the other objects in the object graph. This makes them safe to lift since Dagger shouldn't need to do any wiring for them.

I also reorganized the business logic classes and their modules in a way that I felt made the distinction between global commands and logged-in user commands more clear. This makes it more obvious that the application is divided into two logical pieces centred around the CommandRouter.